requirements.txt
requests
numpy
Pillow
